chore: refactor hooks to common dir
